// src/types/account.ts
export interface IAccountSettingsFormData {
    email: string;
    first_name: string;
    last_name: string;
    phone_number: string;
    company: string;
    job_title: string;
    preferred_contact: string;
    preferred_language: string;
    marketing_opt_in: boolean;
    currentPassword: string;
    newPassword: string;
}